refactor: 도메인 중심 아키텍처 및 전송 계층 분리 (MCP-11)
[Domain Layer]
- graph 패키지를 factcheck로 리네이밍 및 이동
- 내부 노드 구조 정리 및 중첩 제거

[Application Layer]
- mcp_server의 핸들러 및 변환 로직을 application 패키지로 이동
- 비즈니스 로직과 전송 계층의 의존성 분리

[Transport Layer]
- REST(NestJS용)와 MCP(Claude용) 전송 계층 분리
- 라우터 및 스키마 정의 이동 (transport 패키지 신설)

[Test & Scripts]
- 전체 테스트 코드 및 유틸리티 스크립트 마이그레이션
- 정적 분석(Pyright) 및 E2E 테스트 통과 확인
