+++ pyproject.toml
@@ -7,6 +7,9 @@ authors = [
 ]
 readme = "README.md"
 requires-python = ">=3.13,<4.0"
+
+
+
 dependencies = [
     "fastapi (>=0.128.0,<0.129.0)",
     "uvicorn (>=0.40.0,<0.41.0)",
@@ -50,6 +53,7 @@ dev = [
 
 [tool.pyright]
 include = ["src", "tests", "scripts"]
+extraPaths = ["src"]
 pythonVersion = "3.13"
 typeCheckingMode = "standard"
