+++ scripts/test_verification_e2e.py
@@ -6,8 +6,8 @@ load_dotenv()
 
 import asyncio  # noqa: E402
 
-from graph.nodes.verification import verification_node  # noqa: E402
-from graph.state import SentenceState  # noqa: E402
+from factcheck.nodes.verification import verification_node  # noqa: E402
+from factcheck.state import SentenceState  # noqa: E402
 
 """사용법(터미널): PYTHONPATH=src poetry run python scripts/test_verification_e2e.py"""
