--- src/mcp_server/__init__.py
@@ -1,31 +1,0 @@
-"""MCP 프로토콜 계층"""
-
-from mcp_server.errors import JsonRpcErrorCode
-from mcp_server.schemas import (
-    ContentItem,
-    FactcheckArguments,
-    JsonRpcError,
-    JsonRpcErrorResponse,
-    JsonRpcRequest,
-    JsonRpcResponse,
-    JsonRpcSuccessResponse,
-    ToolCallParams,
-    ToolDefinition,
-    ToolsCallResult,
-    ToolsListResult,
-)
-
-__all__ = [
-    "JsonRpcErrorCode",
-    "JsonRpcRequest",
-    "JsonRpcResponse",
-    "JsonRpcSuccessResponse",
-    "JsonRpcErrorResponse",
-    "JsonRpcError",
-    "ToolDefinition",
-    "ToolCallParams",
-    "ToolsListResult",
-    "ToolsCallResult",
-    "ContentItem",
-    "FactcheckArguments",
-]
