--- src/mcp_server/schemas/tools/__init__.py
@@ -1,17 +1,0 @@
-"""도구별 스키마 패키지"""
-
-from mcp_server.schemas.tools.factcheck import (
-    ClaimSentence,
-    FactcheckArguments,
-    FactcheckResult,
-    OpinionSentence,
-    ResultSentence,
-)
-
-__all__ = [
-    "FactcheckArguments",
-    "FactcheckResult",
-    "ClaimSentence",
-    "OpinionSentence",
-    "ResultSentence",
-]
