+++ src/services/llm_spec.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from graph.state import Source
+from factcheck.state import Source
 from services import llm
